+++ src/control_plane/workload/workload_reconcile_action.rs
@@ -5,6 +5,5 @@ pub(crate) enum WorkloadReconcileAction {
     Unchanged,
     Created,
     Started,
-    Restarted,
     Replaced,
 }